feat: React Navigation에서 Expo Router v2로 마이그레이션
- Expo Router를 사용한 파일 기반 라우팅으로 앱 폴더 구조 재구성
- /src/screens의 화면 컴포넌트들을 app 디렉토리로 이동
- 인증 흐름과 컨텍스트 관리 구현
- 로그인/로그아웃 네비게이션 흐름 수정
- 탭과 네비게이션 구조 재구성
- App.tsx를 제거하고 app/_layout.tsx로 대체
- 의존성 및 설정 업데이트
