+++ app/(auth)/(login)/_layout.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+import { Stack } from "expo-router";
+
+export default function LoginLayout() {
+  return (
+    <Stack screenOptions={{ headerShown: false }}>
+      <Stack.Screen name="login" />
+      <Stack.Screen name="signup" />
+    </Stack>
+  );
+}
