+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "pickup",
   "version": "1.0.0",
-  "main": "index.ts",
+  "main": "expo-router/entry",
   "scripts": {
     "start": "expo start --dev-client",
     "android": "expo run:android",
@@ -23,12 +23,14 @@
     "axios": "^1.7.9",
     "eventemitter3": "^5.0.1",
     "expo": "~52.0.26",
-    "expo-constants": "~17.0.6",
+    "expo-constants": "~17.0.8",
     "expo-dev-client": "~5.0.12",
     "expo-device": "~7.0.2",
     "expo-firebase-messaging": "^2.0.0",
     "expo-linear-gradient": "~14.0.2",
+    "expo-linking": "~7.0.5",
     "expo-notifications": "~0.29.13",
+    "expo-router": "~4.0.19",
     "expo-splash-screen": "~0.29.22",
     "expo-status-bar": "~2.0.1",
     "lucide-react-native": "^0.473.0",
@@ -38,9 +40,9 @@
     "react-native-base64": "^0.2.1",
     "react-native-gesture-handler": "^2.22.1",
     "react-native-linear-gradient": "^2.8.3",
-    "react-native-reanimated": "^3.16.7",
-    "react-native-safe-area-context": "^5.1.0",
-    "react-native-screens": "^4.5.0",
+    "react-native-reanimated": "^3.17.1",
+    "react-native-safe-area-context": "4.12.0",
+    "react-native-screens": "~4.4.0",
     "react-native-svg": "15.8.0",
     "react-native-vector-icons": "^10.2.0",
     "react-native-web": "~0.19.13"
