+++ src/components/auth/signup/SignupLink.tsx
@@ -1,22 +1,13 @@
 import { StyleSheet } from "react-native";
 import { Text, TouchableOpacity, View } from "react-native";
-import { useNavigation } from "@react-navigation/native";
-import type { StackNavigationProp } from "@react-navigation/stack";
-
-type RootStackParamList = {
-  Start: undefined;
-  Login: undefined;
-  Signup: undefined;
-};
-
-type NavigationProp = StackNavigationProp<RootStackParamList>;
+import { useRouter } from "expo-router";
 
 const SignupLink = () => {
-  const navigation = useNavigation<NavigationProp>();
+  const router = useRouter();
   return (
     <View style={styles.signupContainer}>
       <Text style={styles.signupText}>아직 계정이 없으신가요?</Text>
-      <TouchableOpacity onPress={() => navigation.navigate("Signup")}>
+      <TouchableOpacity onPress={() => router.push("signup")}>
         <Text style={styles.signupLink}>회원가입하기</Text>
       </TouchableOpacity>
     </View>
