+++ src/utils/auth/handleLogin.ts
@@ -1,14 +1,13 @@
-import { NavigationProp } from "@react-navigation/native";
-import AsyncStorage from "@react-native-async-storage/async-storage";
-import { AxiosError, AxiosInstance } from "axios";
-import { decodeToken } from "./decodeToken";
+import { AxiosError } from "axios";
 import { authEventEmitter } from "../event";
+import { Router } from "expo-router";
+import axiosInstance from "../../api/axios-interceptor";
+import AsyncStorage from "@react-native-async-storage/async-storage";
 
 interface LoginParams {
   email: string;
   password: string;
-  navigation: NavigationProp<any>;
-  axiosInstance: AxiosInstance;
+  router: Router;
   onError?: (message: string) => void;
   onSuccess?: () => void;
   shouldNavigate?: boolean;
@@ -22,35 +21,54 @@ interface LoginResponse {
 export const handleLogin = async ({
   email,
   password,
-  navigation,
-  axiosInstance,
+  router,
   onError,
   onSuccess,
   shouldNavigate = true,
 }: LoginParams): Promise<boolean> => {
   try {
+    console.log("로그인 시도 중:", email);
     const response = await axiosInstance.post<LoginResponse>("/auth/login", {
       email,
       password,
     });
 
     if (response.data?.accessToken) {
-      // 로그인 이벤트 발생
-      authEventEmitter.emit("login", {
-        accessToken: response.data.accessToken,
-        refreshToken: response.data.refreshToken,
-        jti: response.data.jti,
-        navigation,
-        shouldNavigate,
-        callback: onSuccess,
-      });
+      console.log("토큰 받음, 저장 시작");
 
-      return true;
-    } else {
-      console.error("No token in response:", response.data);
-      onError?.("로그인 응답에 토큰이 없습니다");
-      return false;
+      // 저장 전 체크
+      const beforeToken = await AsyncStorage.getItem("accessToken");
+      console.log("저장 전 토큰:", beforeToken ? "있음" : "없음");
+
+      // 토큰 저장
+      await AsyncStorage.setItem("accessToken", response.data.accessToken);
+      await AsyncStorage.setItem("refreshToken", response.data.refreshToken);
+      await AsyncStorage.setItem("jti", response.data.jti);
+      await AsyncStorage.setItem("isLoggedIn", "true"); // 중요!
+
+      // 저장 후 체크
+      const afterToken = await AsyncStorage.getItem("accessToken");
+      const isLoggedInValue = await AsyncStorage.getItem("isLoggedIn");
+      console.log(
+        "저장 후 토큰:",
+        afterToken ? afterToken.substring(0, 10) + "..." : "없음"
+      );
+      console.log("isLoggedIn 값:", isLoggedInValue);
+
+      // 성공 콜백
+      if (onSuccess) {
+        console.log("성공 콜백 호출");
+        onSuccess();
+      }
+
+      // // 네비게이션
+      // if (shouldNavigate) {
+      //   console.log("화면 이동 시도");
+      //   router.replace("/(tabs)");
+      //   console.log("화면 이동 명령 완료");
+      // }
     }
+    return true;
   } catch (error) {
     if (error instanceof AxiosError) {
       console.log("Error details:", {
